[add] user authentication and authorizzation

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from 'src/auth/user/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,35 +15,42 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
   // get all task aswell as filter task with task status
-  getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTask(filterDto);
+  getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTask(filterDto, user);
   }
   // store new task
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
   // get task with id
-  async getTaskById(id: number): Promise<Task> {
-    const taskFound = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    // const taskFound = await this.taskRepository.findOne(id);
+    const taskFound = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     if (!taskFound) {
-      throw new NotFoundException(`Not Available`);
+      throw new NotFoundException(`No Item Found`);
     }
     return taskFound;
   }
 
   // delete task with id
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
-      throw new NotFoundException(`Not Available`);
+      throw new NotFoundException(`No Item Found`);
     }
   }
 
   // update task status with id
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
